test: unified multiple selection actions tests
